Update minimum battery level
Screen was reading as 30% charged when LED indicators were signalling battery depleted. Adjusting minimum battery level to more closely match. I should probably just measure the level of the battery when it's flat - I know how low the LiPo should be allowed to go, but I don't know what the LED indicators are set to as a minimum. Need to remember to bring multimeter.